Tested and Fixed bugs
Tested the application using the files sent by Kevin. Most of the files worked properly and fixed bug where documentTypes which did not have an email address in the appconfig would not send an email properly. Also, fixed the issues, if no emails are attached to the email then default email will be attached. The program should be working as expected.